False Positives

False Positives

MLMetrics.false_positives — Function.

false_positives(target, output, [encoding]) -> Int

Return 1 if target is considered a negative label and output is considered a positive label (according to encoding). Return 0 otherwise. This is also known as type 1 error.

false_positives(targets::AbstractArray, outputs::AbstractArray, [encoding]) -> Union{Int, Dict}

Count the total number of false positives in outputs by comparing each element against the corresponding value in targets (according to encoding). Both parameters are expected to be vectors of some form, which means they are allowed to be row-vectors (or row-matrices).

The optional parameter encoding serves as specifcation of the existing labels and their interpretation (e.g. what constitutes as positive or negative, how many classes exist, etc). It can either be an object from the namespace LabelEnc, or a vector of labels. If omitted, the appropriate encoding will be inferred from the types and/or values of targets and outputs. Note that omitting the encoding can cause performance penalties, which may include a lack of type stability.