True Positives

True Positives

MLMetrics.true_positives — Function.

true_positives(targets, outputs, [encoding]) -> Union{Int, Dict}

Count how many positive predicted outcomes in outputs are also marked as positive outcomes in targets. Which value denotes "positive" depends on the given (or inferred) encoding. Typically both parameters are arrays of some form, (e.g. vectors or row-vectors), but its also possible to provide a single obseration as "scalar" value.

The return value of the function depends on the number of labels in the given encoding. In case the encoding is binary (i.e. it has exactly 2 labels), a single integer value is returned. Otherwise, the function will compute a separate result for each individual label, where that label is treated as "positive" and the other labels are treated as "negative". These results are then returned as a single dictionary with an entry for each label.

If encoding is omitted, the appropriate MLLabelUtils.LabelEncoding will be inferred from the types and/or values of targets and outputs. Note that omitting the encoding can cause performance penalties, which may include a lack of return-type inference.

Arguments

targets: Either an array of multiple ground truths $\mathbf{y}$, or a single ground truth $y$.

outputs: Either an array of multiple predicted outputs $\mathbf{\hat{y}}$, or a single prediction $\hat{y}$.

encoding: Optional. Specifies the possible values in targets and outputs and their interpretation (e.g. what constitutes as a positive or negative label, how many labels exist, etc). It can either be an object from the namespace LabelEnc, or a vector of labels.

See also

predicted_positive, condition_positive, true_positive_rate

Examples

julia> true_positives(1, 1, LabelEnc.ZeroOne()) # single observation
1

julia> true_positives([1,0,1,1,0], [1,1,1,0,0]) # multiple observations
2

julia> true_positives([:a,:a,:b,:b,:c,:c], [:a,:b,:b,:b,:a,:a]) # multi-class
Dict{Symbol,Int64} with 3 entries:
  :a => 1
  :b => 2
  :c => 0