F-Score

F-Score

MLMetrics.f_score — Function.

f_score(targets, outputs, [encoding], [avgmode], [β = 1]) -> Float64

Compute the F-score for the outputs given the targets. The F-score is a measure for accessing the quality of binary predictor by considering both recall and the precision.

julia> recall([1,0,0,1,1], [1,0,0,0,1])
0.6666666666666666

julia> precision_score([1,0,0,1,1], [1,0,0,0,1])
1.0

julia> f_score([1,0,0,1,1], [1,0,0,0,1])
0.8

The parameter β can be used to balance the importance of recall vs precision. The default β = 1 corresponds to the harmonic mean. A value of β > 1 weighs recall higher than precision, while a value of β < 1 weighs recall lower than precision.

julia> f_score([1,0,0,1,1], [1,0,0,0,1], 2)
0.7142857142857143

julia> f_score([1,0,0,1,1], [1,0,0,0,1], 0.5)
0.9090909090909091

The optional parameter encoding serves as specifcation of the existing labels and their interpretation (e.g. what constitutes as positive or negative, how many classes exist, etc). It can either be an object from the namespace LabelEnc, or a vector of labels. If omitted, the appropriate encoding will be inferred from the types and/or values of targets and outputs. Note that omitting the encoding can cause performance penalties, which may include a lack of type stability.

julia> f_score([1,0,0,1,1], [1,-1,-1,-1,1], LabelEnc.FuzzyBinary())
0.8

MLMetrics.f1_score — Function.

f1_score(targets, outputs, [encoding], [avgmode])

Same as f_score, but with β fixed to 1.