Positive Likelihood Ratio

Positive Likelihood Ratio

MLMetrics.positive_likelihood_ratio — Function.

positive_likelihood_ratio(targets, outputs, [encoding], [avgmode])

Compute the positive likelihood ratio for the given outputs and targets. It is a useful meassure for assessing the quality of a diagnostic test and is defined as sensitivity / (1 - specificity). This can also be written as true_positive_rate / false_positive_rate.

julia> positive_likelihood_ratio([0,1,1,0,1], [1,1,1,0,1])
2.0

julia> positive_likelihood_ratio([-1,1,1,-1,1], [1,1,1,-1,1])
2.0

The optional parameter encoding serves as specifcation of the existing labels and their interpretation (e.g. what constitutes as positive or negative, how many classes exist, etc). It can either be an object from the namespace LabelEnc, or a vector of labels. If omitted, the appropriate encoding will be inferred from the types and/or values of targets and outputs. Note that omitting the encoding can cause performance penalties, which may include a lack of type stability.

julia> positive_likelihood_ratio([:b,:b,:a,:c,:c], [:a,:c,:b,:c,:c], LabelEnc.OneVsRest(:c))
3.0

The optional (keyword) parameter avgmode can be used to specify if and how class-specific results should be aggregated. This is mainly useful if there are more than two classes. Typical values are :none (default), :micro for micro averaging, or :macro for macro averaging. It is also possible to specify avgmode as a type-stable positional argument using an object from the AvgMode namespace.

julia> positive_likelihood_ratio([:b,:b,:a,:c,:c], [:a,:c,:b,:c,:c]) # avgmode=:none
Dict{Symbol,Float64} with 3 entries:
  :a => 0.0
  :b => 0.0
  :c => 3.0

julia> positive_likelihood_ratio([:b,:b,:a,:c,:c], [:a,:c,:b,:c,:c], avgmode=:micro)
1.3333333333333335

julia> positive_likelihood_ratio([:b,:b,:a,:c,:c], [:a,:c,:b,:c,:c], avgmode=:macro)
1.090909090909091