Prevalence

Prevalence

MLMetrics.prevalence — Function.

prevalence(targets, outputs, [encoding], [avgmode])

Return the fraction of positive observations in targets. What constitutes as positive depends on encoding.

julia> prevalence([0,1,1,0,1], [1,1,1,0,1])
0.6

julia> prevalence([-1,1,1,-1,1], [1,1,1,-1,1])
0.6

The optional parameter encoding serves as specifcation of the existing labels and their interpretation (e.g. what constitutes as positive or negative, how many classes exist, etc). It can either be an object from the namespace LabelEnc, or a vector of labels. If omitted, the appropriate encoding will be inferred from the types and/or values of targets and outputs. Note that omitting the encoding can cause performance penalties, which may include a lack of type stability.

julia> prevalence([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c], LabelEnc.OneVsRest(:c))
0.4

The optional (keyword) parameter avgmode can be used to specify if and how class-specific results should be aggregated (default: none). Note though that this functionality is mainly provided for interface consistency.

julia> prevalence([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c]) # avgmode=:none
Dict{Symbol,Float64} with 3 entries:
  :a => 0.4
  :b => 0.2
  :c => 0.4

julia> prevalence([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c], avgmode=:micro)
0.3333333333333333