Accuracy

Accuracy

MLMetrics.accuracy — Function.

accuracy(targets, outputs, [encoding]; [normalize = true]) -> Float64

Compute the classification accuracy for the outputs given the targets. If normalize is true, the fraction of correctly classified observations in outputs is returned (according to encoding). Otherwise the total number is returned.

julia> accuracy([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c])
0.6

julia> accuracy([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c], normalize=false)
3.0

The optional parameter encoding serves as specifcation of the existing labels and their interpretation (e.g. what constitutes as positive or negative, how many classes exist, etc). It can either be an object from the namespace LabelEnc, or a vector of labels. If omitted, the appropriate encoding will be inferred from the types and/or values of targets and outputs. Note that omitting the encoding can cause performance penalties, which may include a lack of type stability.

julia> accuracy([1,0,0,1,1], [1,-1,-1,-1,1], LabelEnc.FuzzyBinary())
0.8