Positive Predictive Value

Positive Predictive Value

MLMetrics.positive_predictive_value — Function.

positive_predictive_value(targets, outputs, [encoding], [avgmode])

Return the fraction of positive predicted outcomes in outputs that are true positives according to the correspondig targets. This is also known as "precision" (alias precision_score).

julia> precision_score([0,1,1,0,1], [1,1,1,0,1])
0.75

julia> precision_score([-1,1,1,-1,1], [1,1,1,-1,1])
0.75

The optional parameter encoding serves as specifcation of the existing labels and their interpretation (e.g. what constitutes as positive or negative, how many classes exist, etc). It can either be an object from the namespace LabelEnc, or a vector of labels. If omitted, the appropriate encoding will be inferred from the types and/or values of targets and outputs. Note that omitting the encoding can cause performance penalties, which may include a lack of type stability.

julia> precision_score([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c], LabelEnc.OneVsRest(:c))
0.6666666666666666

The optional (keyword) parameter avgmode can be used to specify if and how class-specific results should be aggregated. This is mainly useful if there are more than two classes. Typical values are :none (default), :micro for micro averaging, or :macro for macro averaging. It is also possible to specify avgmode as a type-stable positional argument using an object from the AvgMode namespace.

julia> precision_score([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c]) # avgmode=:none
Dict{Symbol,Float64} with 3 entries:
  :a => 1.0
  :b => 0.0
  :c => 0.666667

julia> precision_score([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c], avgmode=:micro)
0.6

julia> precision_score([:a,:b,:a,:c,:c], [:a,:c,:b,:c,:c], avgmode=:macro)
0.5555555555555555